% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_licenses.R
\name{get_licenses}
\alias{get_licenses}
\title{retrieves licenses of all attached and loaded packages}
\usage{
get_licenses()
}
\value{
A data frame of the names, versions, licenses and repositories
    of the attached and loaded packages
}
\description{
\code{get_licenses} retrieves the versions, licenses and repositories of
all the attached and loaded R packages in an R session.
}
\examples{
get_licenses()
}
