% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcutils.R
\docType{package}
\name{stcutils}
\alias{stcutils}
\alias{stcutils-package}
\title{stcutils: A package to facilitate adherence to Statistics Canada's R coding
best practices, in particular, in relation to reproducibility and
visualization styling.}
\description{
stcutils: A package to facilitate adherence to Statistics Canada's R coding
best practices, in particular, in relation to reproducibility and
visualization styling.
}
